Replace FSM signal post_transition against transition_change_notification function
